Switch shading to single precision, part 4
Convert between single and double precision at the interface between the renderer's core and the modeling entities (BSDF, EDF, BSSRDF).

+++ src/appleseed/renderer/kernel/shading/oslshadergroupexec.h
@@ -99,16 +99,16 @@ class OSLShaderGroupExec
     void execute_bump(
         const ShaderGroup&              shader_group,
         const ShadingPoint&             shading_point,
-        const foundation::Vector2d&     s) const;
+        const foundation::Vector2f&     s) const;
 
     void choose_subsurface_normal(
         const ShadingPoint&             shading_point,
         const void*                     bssrdf_data,
-        const double                    s) const;
+        const float                     s) const;
 
     foundation::Color3f execute_background(
         const ShaderGroup&              shader_group,
-        const foundation::Vector3d&     outgoing) const;
+        const foundation::Vector3f&     outgoing) const;
 
     void do_execute(
         const ShaderGroup&              shader_group,
